AG-17862 Strip example generator harness from markdown examples
Extract stripOutExampleGeneratorCode from CodeViewer into a standalone,
React/SCSS-free module so it can run in the build-time markdown
serializer. The .md docs pages now strip the same injected harness the
on-page code viewer does (test-id setup, console logging, teardown,
theme switcher, AI-token redaction) instead of only the test-id block.
