+++ documentation/ag-grid-docs/src/components/example-runner/components/CodeViewer.tsx
@@ -1,12 +1,6 @@
 import type { InternalFramework } from '@ag-grid-types';
 import Code from '@ag-website-shared/components/code/Code';
 import { Icon } from '@ag-website-shared/components/icon/Icon';
-import {
-    CONSOLE_LOG_REGEX,
-    DARK_INTEGRATED_REGEX,
-    TEAR_DOWN_REGEX,
-    TEST_ID_REGEX,
-} from '@ag-website-shared/utils/extraCodeSnippets';
 import type { FileContents } from '@components/example-generator/types';
 import { doOnEnter } from '@utils/doOnEnter';
 import classnames from 'classnames';
@@ -14,6 +8,7 @@ import { useEffect, useState } from 'react';
 
 import { CodeOptions } from './CodeOptions';
 import styles from './CodeViewer.module.scss';
+import { stripOutExampleGeneratorCode } from './stripOutExampleGeneratorCode';
 
 const ExtensionMap = {
     sh: 'bash',
@@ -21,34 +16,6 @@ const ExtensionMap = {
     tsx: 'jsx',
     json: 'js',
 };
-
-const AI_API_TOKEN_REGEX = /(const AI_API_TOKEN\s*=\s*)(['"`])[^'"`]+\2/g;
-
-export function stripOutExampleGeneratorCode(files: FileContents) {
-    const mainFiles = ['main.js', 'main.ts', 'index.tsx', 'index.jsx', 'app.component.ts'];
-
-    mainFiles.forEach((mainFile) => {
-        if (files[mainFile]) {
-            // hide integrated theme switcher
-            files[mainFile] =
-                files[mainFile]?.replace(DARK_INTEGRATED_REGEX, '').replace(CONSOLE_LOG_REGEX, '').trim() + '\n';
-
-            // Hide the example tear down code use in Documentation tests
-            files[mainFile] = files[mainFile]?.replace(TEAR_DOWN_REGEX, '').trim() + '\n';
-
-            // Hide the test id setup code
-            files[mainFile] = files[mainFile]?.replace(TEST_ID_REGEX, '').trim() + '\n';
-        }
-    });
-
-    // Strip AI API token values from all files before displaying in the code viewer
-    // and show as redacted. If empty, it will show as empty string.
-    for (const fileName of Object.keys(files)) {
-        if (typeof files[fileName] === 'string') {
-            files[fileName] = files[fileName]!.replace(AI_API_TOKEN_REGEX, '$1$2<TOKEN_REDACTED>$2');
-        }
-    }
-}
 
 /**
  * This renders the code viewer in the example runner.
